pyproject.toml
[tool.poetry]
name = "euvd-mcp"
version = "1.2.0"
description = ""
authors = ["Duarte Dias <me@duartedias.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.14"
fastmcp = "3.2.1"
httpx = "^0.28.1"
pydantic-settings = "2.13.1"
cachetools = "^7.0"
uvicorn = {extras = ["standard"], version = "0.44.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-cov = "7.1.0"
pytest-mock = "3.15.1"
pytest-asyncio = "1.3.0"
pytest-httpx = "0.36.0"
types-cachetools = "^6.2.0.20260408"
isort = "8.0.1"
flake8 = "7.3.0"
mypy = "1.20.0"
pylint = "4.0.5"
ruff = "0.15.9"
bandit = "1.9.4"
safety = "3.7.0"
pip-audit = "2.10.0"
black = "26.3.1"
pre-commit = "4.5.1"
pydocstyle = "6.3.0"

[tool.flake8]
max-line-length = 127
exclude = [".git", "__pycache__", ".venv", "venv", "build", "dist", ".tox", ".eggs", "*.egg"]
ignore = ["E203", "E266", "E501", "W503"]
max-complexity = 10

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = false
ignore_missing_imports = true
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "euvd_mcp.tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "euvd_mcp.main"
disable_error_code = ["untyped-decorator"]

[tool.pytest.ini_options]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["euvd_mcp/tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests"
]
addopts = "-v --strict-markers --tb=short"
asyncio_mode = "auto"

[tool.pydocstyle]
convention = "google"
add_ignore = ["D212", "D415"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
